requirements.txt
Django>=5.1.6
djangorestframework>=3.14.0
pdfplumber>=0.10.0
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0
weasyprint>=60.0
python-dotenv>=1.0.0
